requirements.txt
# Production dependencies
aiogram==3.12.0
SQLAlchemy==2.0.32
aiosqlite==0.20.0
pydantic-settings==2.3.4

# Development dependencies (install with: pip install -e ".[dev]")
# pytest==8.3.2
# pytest-cov==4.1.0
# black==24.8.0
# ruff==0.6.9
# mypy==1.11.2
# pre-commit==3.6.0
# safety==2.3.5
# bandit==1.7.5
